# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/v1/admin/goodsType/addGoodsType": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "添加商品种类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品名称",
                        "name": "goods_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品规格 1.盒 2.瓶 3.支",
                        "name": "goods_specs",
                        "in": "formData"
                    },
                    {
                        "type": "float",
                        "description": "商品成本价",
                        "name": "goods_unitprince",
                        "in": "formData"
                    },
                    {
                        "type": "float",
                        "description": "商品销售价",
                        "name": "goods_prince",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "商品图片",
                        "name": "goods_image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "生产批号",
                        "name": "goods_batch_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "生产日期",
                        "name": "goods_date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "商品状态 1.下架  2.在售",
                        "name": "goods_state",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/deleteGoodsType": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "下架商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/deleteGoodsTypes": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "批量下架商品",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.DeleteIds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/queryAllGoods": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "查询商品种类ID和商品名（支持模糊查询）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品名称",
                        "name": "goods_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/queryByGoodsTypeID": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "查看商品种类详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品种类ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/queryGoodsTypesByLimitOffset": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "分页查询商品种类(默认前100条) 并返回总记录数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/goodsType/updateGoodsType": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品种类"
                ],
                "summary": "修改商品种类信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "goods_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品名称",
                        "name": "goods_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品规格 1.盒 2.瓶 3.支",
                        "name": "goods_specs",
                        "in": "formData"
                    },
                    {
                        "type": "float",
                        "description": "商品成本价",
                        "name": "goods_unitprince",
                        "in": "formData"
                    },
                    {
                        "type": "float",
                        "description": "商品销售价",
                        "name": "goods_prince",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "商品图片",
                        "name": "goods_image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "生产批号",
                        "name": "goods_batch_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "生产日期",
                        "name": "goods_date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "商品状态 1.下架  2.在售",
                        "name": "goods_state",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.DeleteIds": {
            "type": "object",
            "properties": {
                "ids": {
                    "description": "ids",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "entity.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}